requirements.txt
pandas
nltk
matplotlib
newspaper3k
duckduckgo-search
ddgs
keybert
transformers
torch
scikit-learn
lxml_html_clean
yfinance

